+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+## [0.1.79] - 2024-04-25
+Implemented a new functionality to send email notifications regarding the initiation of EMR cluster spin-up during the start of datapull jobs, covering both success and failure scenarios.
+### Changed
+- api/src/main/java/com/homeaway/datapullclient/process/DataPullTask.java
+- api/src/main/java/com/homeaway/datapullclient/process/DataPullRequestProcessor.java
+- api/src/main/java/com/homeaway/datapullclient/config/DataPullClientConfig.java
+### Added
+- api/src/main/java/com/homeaway/datapullclient/config/SESProperties.java
+- api/src/main/java/com/homeaway/datapullclient/config/SMTPProperties.java
+- api/src/main/java/com/homeaway/datapullclient/utils/EmailNotification.java
+
 ## [0.1.78] - 2024-04-22
 Mandating the Linux flavor of the docker image that's been used in ECR.
 ### Changed
@@ -14,7 +25,6 @@ Mandating the Linux flavor of the docker image that's been used in ECR.
 Fixing the null subnetId issue by adding  StringUtils.isNotBlank(dataPullProperties.getApplicationSubnet3())
 ### Changed
 - api/src/main/java/com/homeaway/datapullclient/process/DataPullTask.java
-
 
 ## [0.1.76] - 2024-01-08
 Fixing the issue when the key is not provided by the user. and reverting chnages w.r.t insertinto hive table for partitioned table.